Crate mahowald_invariant

Computes algebraic Mahowald invariants (aka algebraic root invariants).

Sample output (with Max k = 7):

M({basis element}) = {mahowald_invariant}[ mod {indeterminacy}]
M(x_(0, 0, 0)) = x_(0, 0, 0)
M(x_(1, 1, 0)) = x_(1, 2, 0)
M(x_(2, 2, 0)) = x_(2, 4, 0)
M(x_(1, 2, 0)) = x_(1, 4, 0)
M(x_(3, 3, 0)) = x_(3, 6, 0)
M(x_(2, 4, 0)) = x_(2, 8, 0)
M(x_(1, 4, 0)) = x_(1, 8, 0)
M(x_(2, 5, 0)) = x_(2, 10, 0)
M(x_(3, 6, 0)) = x_(3, 12, 0)

Here is a brief overview of what this example computes. For details and beyond, see for instance “The root invariant in homotopy theory” or “The Bredon-Löffler conjecture” (where the latter also contains machine computations similar to what this example does). In the following, we abbreviate Ext^{s,t}_A(-, F_2) as Ext^{s,t}(-).

Let M_k be the cohomology of RP_-k_inf. There is an isomorphism Ext^{s, t}(F_2) ~ lim_k Ext^{s, t-1}(M_k) induced by the (-1)-cell S^{-1} -> RP_-k_inf at each level. Let x be a class in Ext^{s, t}(F_2). Then there is a minimal k such that its image in Ext^{s, t-1}(M_k) is non-trivial. Using the long exact sequence induced by the (co)fiber sequence S^{-k} -> RP_-k_inf -> RP_{-k+1}_inf on the level of Ext, that image can be lifted to a class M(x) in Ext^{s, t + k - 1}, which is (a representative for) the (algebraic) Mahowald invariant of x.

This script computes these lifts (and their indeterminacy) by resolving F_2 resp. M_ks and constructing ResolutionHomomorphisms corresponding to the bottom and (-1)-cells. Given Max k, it will print Mahowald invariants of the F_2-basis elements of Ext^{*,*}(F_2) that are detected in Ext^{*,*}(M_k) for the first time for some k <= Max k.